+++ tee-worker/identity/litentry/core/native-task/receiver/src/trusted_call_handlers/request_vc_handler.rs
